Move blocks into separate files
