+++ apps/src/music/blockly/blockTypes.js
@@ -1,0 +1,11 @@
+export const BlockTypes = {
+  WHEN_RUN: 'when_run',
+  WHEN_TRIGGER: 'when_trigger',
+  PLAY_SOUND: 'play_sound',
+  PLAY_SOUND_NEXT_MEASURE: 'play_sound_next_measure',
+  LOOP_FROM_TO: 'loop_from_to',
+  IF_EVEN_THEN: 'if_even_then',
+  NUMBER: 'number',
+  VARIABLES_GET: 'variables_get',
+  VARIABLES_SET: 'variables_set'
+};
